requirements.txt
azure-functions
azure-cosmos
geopy
cryptography
